requirements.txt
# Core Libraries
pillow>=10.0.0
pytesseract>=0.3.10
pdfplumber>=0.10.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0

# AI & NLP Libraries
transformers>=4.30.0
torch>=2.0.0
spacy>=3.5.0

# Hebrew NLP Support
# Install Hebrew model after: python -m spacy download xx_ent_wiki_sm

# Additional utilities
numpy>=1.24.0
regex>=2023.0.0
tqdm>=4.65.0
